services/orchestrator/requirements.txt
fastapi==0.115.0
uvicorn==0.30.5
redis==5.0.1
livekit-api==1.0.5
prometheus-client==0.20.0
python-multipart==0.0.6
